requirements.txt
sentence-transformers==2.2.2
chromadb==0.4.22
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
pymupdf==1.23.26
numpy==1.24.3
tqdm==4.66.2
argparse==1.4.0
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
pdfplumber==0.10.3
opencv-python==4.9.0.80
matplotlib==3.8.2
jupyter==1.0.0